+++ neo/game/script/Script_Thread.cpp
@@ -610,7 +610,7 @@ void idThread::KillThread( const char *name ) {
 	if ( ptr ) {
 		len = ptr - name;
 	} else {
-		len = strlen( name );
+		len = (int)strlen( name );
 	}
 
 	// kill only those threads whose name matches name
@@ -1615,7 +1615,7 @@ idThread::Event_StrLen
 void idThread::Event_StrLen( const char *string ) {
 	int len;
 
-	len = strlen( string );
+	len = (int)strlen( string );
 	idThread::ReturnInt( len );
 }
 
@@ -1632,7 +1632,7 @@ void idThread::Event_StrLeft( const char *string, int num ) {
 		return;
 	}
 
-	len = strlen( string );
+	len = (int)strlen( string );
 	if ( len < num ) {
 		idThread::ReturnString( string );
 		return;
@@ -1655,7 +1655,7 @@ void idThread::Event_StrRight( const char *string, int num ) {
 		return;
 	}
 
-	len = strlen( string );
+	len = (int)strlen( string );
 	if ( len < num ) {
 		idThread::ReturnString( string );
 		return;
@@ -1677,7 +1677,7 @@ void idThread::Event_StrSkip( const char *string, int num ) {
 		return;
 	}
 
-	len = strlen( string );
+	len = (int)strlen( string );
 	if ( len < num ) {
 		idThread::ReturnString( "" );
 		return;
@@ -1702,7 +1702,7 @@ void idThread::Event_StrMid( const char *string, int start, int num ) {
 	if ( start < 0 ) {
 		start = 0;
 	}
-	len = strlen( string );
+	len = (int)strlen( string );
 	if ( start > len ) {
 		start = len;
 	}
